septidon-integration: reorganize module hierarchy
