+++ src/poseidon.rs
@@ -13,6 +13,9 @@ use halo2_proofs::{
 mod pow5;
 pub use pow5::{Pow5Chip, Pow5Config, StateWord, Var};
 
+mod septidon;
+pub use septidon::SeptidonChip;
+
 pub mod primitives;
 use primitives::{Absorbing, ConstantLength, Domain, Spec, SpongeMode, Squeezing, State};
 use std::fmt::Debug as DebugT;
