+++ src/poseidon/septidon/transition_round.rs
@@ -1,7 +1,7 @@
+use super::params;
+use super::params::{mds, round_constant};
 use super::state::Cell;
-use crate::septidon::params;
-use crate::septidon::params::{mds, round_constant};
-use crate::septidon::util::{join_values, matmul, split_values};
+use super::util::{join_values, matmul, split_values};
 use halo2_proofs::circuit::{Region, Value};
 use halo2_proofs::halo2curves::bn256::Fr as F;
 use halo2_proofs::plonk::{Advice, Column, ConstraintSystem, Constraints, Error, Expression};
